backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
asyncpg==0.29.0
databases==0.8.0
sqlalchemy==2.0.25

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Supabase
supabase==2.3.4
gotrue==2.4.2

# Storage & Files
boto3==1.34.34
Pillow==10.2.0

# Email & SMS
sendgrid==6.11.0
twilio==8.12.0

# Payment
stripe==7.12.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
httpx==0.26.0
aiofiles==23.2.1

# Monitoring & Logging
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
